refactor(server): preserve readability thresholds
